disc_bot/MCServer_bot/requirements.txt
discord.py
python-dotenv
wakeonlan
paramiko